+++ sync-basalam/includes/Actions/Controller/ProductActions/DetectionProductCategories.php
@@ -4,6 +4,7 @@ namespace SyncBasalam\Actions\Controller\ProductActions;
 
 use SyncBasalam\Services\Products\GetCategoryId;
 use SyncBasalam\Actions\Controller\ActionController;
+use SyncBasalam\Admin\Settings\SettingsConfig;
 
 defined('ABSPATH') || exit;
 
@@ -15,7 +16,16 @@ class DetectionProductCategories extends ActionController
             wp_send_json_error(['message' => 'عنوان محصول ارسال نشده است.'], 400);
         }
 
-        $productTitle = mb_substr(sanitize_text_field(wp_unslash($_POST['productTitle'])), 0, 120);
+        $productTitle = sanitize_text_field(wp_unslash($_POST['productTitle']));
+
+        $prefix = syncBasalamSettings()->getSettings(SettingsConfig::PRODUCT_PREFIX_TITLE);
+        $suffix = syncBasalamSettings()->getSettings(SettingsConfig::PRODUCT_SUFFIX_TITLE);
+
+        if (!empty($prefix)) $productTitle = $prefix . ' ' . $productTitle;
+        if (!empty($suffix)) $productTitle = $productTitle . ' ' . $suffix;
+
+        $productTitle = mb_substr($productTitle, 0, 120);
+
         $categoryIds = GetCategoryId::getCategoryIdFromBasalam($productTitle, 'all');
 
         if ($categoryIds && is_array($categoryIds)) {